Remove iidm network dependency in DefaultDiagramStyleProvider, because it is subclassed by Plasma project, that do not rely on IIDM.
