+++ single-line-diagram-core/src/test/java/com/powsybl/sld/TestCase11SubstationGraph.java
@@ -247,6 +247,6 @@ public class TestCase11SubstationGraph extends AbstractTestCase {
 
         compareMetadata(diagram, layoutParameters, "/substDiag_metadata.json",
                 new DefaultDiagramInitialValueProvider(network),
-                new DefaultDiagramStyleProvider(network));
+                new DefaultDiagramStyleProvider());
     }
 }
