+++ single-line-diagram-view-app/src/main/java/com/powsybl/sld/view/app/AbstractSingleLineDiagramViewer.java
@@ -1108,13 +1108,12 @@ public abstract class AbstractSingleLineDiagramViewer extends Application implem
     }
 
     private void initStylesProvider() {
-        styles.put("Default", null);
+        styles.put("Default", new DefaultDiagramStyleProvider());
         styles.put("Nominal voltage", null);
         styles.put("Topology", null);
     }
 
     private void updateStylesProvider(Network network) {
-        styles.put("Default", new DefaultDiagramStyleProvider(network));
         styles.put("Nominal voltage", new NominalVoltageDiagramStyleProvider(network));
         styles.put("Topology", new TopologicalStyleProvider(network));
     }
